% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oocRPCAWrappers.R
\name{oocPCA_base}
\alias{oocPCA_base}
\title{Prepare the oocPCA result object}
\usage{
oocPCA_base(base_outname, k)
}
\arguments{
\item{base_outname}{The basename for the matrices U, V, and S from the fastpca.xx call}

\item{k}{Rank of the transformation.}
}
\value{
A list containing the decomposition.
}
\description{
This is an internal function that formats the resulting oocPCA object so that it can be handled appropriately by R
}
