% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oocRPCAWrappers.R
\name{oocPCA_csv2binary}
\alias{oocPCA_csv2binary}
\title{Convert CSV to binary format}
\usage{
oocPCA_csv2binary(inputFile, outputFile)
}
\arguments{
\item{inputFile}{file containing matrix to convert}

\item{outputFile}{name of output file}
}
\description{
The the computation time of an out-of-core algorithm is typically determined by the cost of loading 
the matrix from the slow memory (i.e. the hard disk).  It is much faster to load a matrix from a binary format
than CSV, because the former represents it in much more compressed format.  This function converts 
a CSV to the compressed binary format.
}
